+++ clang-tools-extra/test/clang-tidy/checkers/daedalean-preprocessing-directives-2.hh
@@ -1,0 +1,5 @@
+// RUN: %check_clang_tidy %s daedalean-preprocessing-directives %t
+#pragma once
+
+#define A
+// CHECK-MESSAGES: :[[@LINE-1]]:9: warning: Preprocessor directives must not be used [daedalean-preprocessing-directives]